consisten use of testcontainers across tests
